Add new exception that is raised when table has no PKs
Apparently in SQLAnywhere, you can have a table that has no primary keys, and when it tries to load a table with no primary keys (like through reflection), then in SQLAnyDialectget_pk_constraint(), it threw an uncaught TypeError because in the first sql statement, it tries to get the primary keys, which returns None, and then in the second sql statement, it tries to use the nonexistent result. This commit adds a new Exception class, SQLAnyNoPrimaryKeyError, which is thrown if this condition arises, because it makes no sense to be using something like sqlalchemy with tables that have no primary keys. I'm not sure if i'm supposed to be using sqlalchemy.exec.DBAPIError, but there is no documentation or examples on when or how that is thrown, and through the code, it seems that the Engine takes care of that, but its not expecting a reflection call to fail, so its not doing a "try except Exception: raise DBAPIError()" like it does in other places.